change damage resource from Gun to Ammo

+++ src/main/kotlin/com/gmail/vincent031525/extractionshooter/datamap/AmmoStats.kt
@@ -3,13 +3,13 @@ package com.gmail.vincent031525.extractionshooter.datamap
 import com.mojang.serialization.Codec
 import com.mojang.serialization.codecs.RecordCodecBuilder
 
-data class AmmoStats(val damage: Int) {
+data class AmmoStats(val damage: Float) {
 
     companion object {
         val CODEC: Codec<AmmoStats> =
             RecordCodecBuilder.create { instance ->
                 instance.group(
-                    Codec.INT.fieldOf("damage").forGetter { it.damage },
+                    Codec.FLOAT.fieldOf("damage").forGetter { it.damage },
                 ).apply(instance, ::AmmoStats)
             }
     }
